Fix entrypoint to interpret arguments correctly
The entrypoint.sh script had a bug where it interpreted its arguments
correctly. Giving it a single file to validate did not work (two or more
files worked as expected). This has been fixed.

The cause was that both an ENTRYPOINT and a CMD were specified in the
Dockerfile, which causes this behavior.